digital_wallet/app_user/views_apis/update_user.py
from rest_framework.generics import RetrieveUpdateAPIView
from ..models import User
from serializer import UserSerializer

class UpdateUser(RetrieveUpdateAPIView):
    queryset = User.objects.all()
    serializer_class = UserSerializer
    lookup_fields = ['id']
